changet timeInfo.Oprating type to float64 because unmarshal error

+++ response.go
@@ -21,7 +21,7 @@ type timeInfo struct {
 	DateStart        time.Time `json:"date_start"`
 	DateFinish       time.Time `json:"date_finish"`
 	OperatingResetAt int64     `json:"operating_reset_at"`
-	Operating        int       `json:"operating"`
+	Operating        float64   `json:"operating"`
 }
 
 type errInfo struct {
